+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ## [Unreleased]
 
+## [5.13.0] - 2026-02-03
+
 ### Added
 
 - A builtin extension for `telescope` - see `:h compile-mode-telescope-extension`
@@ -548,6 +550,7 @@ Also includes several documentation changes.
     API functions
 
 [unreleased]: https://github.com/ej-shafran/compile-mode.nvim/compare/latest...nightly
+[5.13.0]: https://github.com/ej-shafran/compile-mode.nvim/compare/v5.12.0...v5.13.0
 [5.12.0]: https://github.com/ej-shafran/compile-mode.nvim/compare/v5.11.0...v5.12.0
 [5.11.0]: https://github.com/ej-shafran/compile-mode.nvim/compare/v5.10.2...v5.11.0
 [5.10.2]: https://github.com/ej-shafran/compile-mode.nvim/compare/v5.10.1...v5.10.2